+++ changelog.md
@@ -28,6 +28,9 @@ This version is not yet released. If you are reading this on the website, then t
   - Getting the last row is a very common operation
 - Add [`&ep`](https://uiua.org/docs/&ep) and [`&epf`](https://uiua.org/docs/&epf) system functions for easier printing to stderr
 - Change [`backward 𝄈`](https://uiua.org/docs/backward)'s glyph to `𝄈`. Code using `˜` will continue to work and will be formatted as `𝄈`.
+- Add the experimental [`or ∨`](https://uiua.org/docs/or) function
+  - It has a useful reduction identity
+  - It is also GCD
 - Add experimental [inline macros](https://www.uiua.org/docs/experimental#inline-macros)
   - Deprecate the experimental `stringify` and `signature` modifiers in favor of inline code macros
 - Add experimental [`binary`](https://uiua.org/docs/binary) function, which encodes and decodes arrays into a compact binary representation